split starter from stoper
