+++ benchmarks/benchmarks/pipebenchmark.py
@@ -91,8 +91,8 @@ class WriteWorker(Process):
             except Exception:
                 self.__errors += 1
             else:
-                etime = time.time()
                 self.__counter += 1
+            etime = time.time()
         self.__qresult.put(
             utils.Result(self.__wid, self.__counter, etime - stime,
                          self.__errors))
@@ -148,8 +148,8 @@ class ReadWorker(Process):
             except Exception:
                 self.__errors += 1
             else:
-                etime = time.time()
                 self.__counter += 1
+            etime = time.time()
         self.__qresult.put(
             utils.Result(self.__wid, self.__counter, etime - stime,
                          self.__errors))
